pyproject.toml
[tool.poetry]
name = "parchive"
version = "0.1.0"
description = "A CLI TUI for managing URLs with names and episodes"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.13"
rich = "^13.7.0"
typer = "^0.15.4"
openai = "^1.79.0"
requests = "^2.32.3"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.11.1"

[build-system]
requires = ["poetry.core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
parchive = "app.main:app" 

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"